Split out meta selectors
